Long을 int로 변경(idolId(null)은 idolId(0)으로 변경)

+++ subscription-service/src/main/java/com/bit/docker/subscriptionservice/dto/GroupSubscriptionCreateRequest.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class GroupSubscriptionCreateRequest {
 
     @NotNull
-    private Long groupId;
+    private int groupId;
 
     private boolean autoRenew;
 }
